roles permissions step-2

+++ app/Http/Controllers/Admin/UsersController.php
@@ -5,8 +5,10 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rules\Password;
-use Illuminate\Support\Facades\Hash;
 
+use Illuminate\Support\Facades\Hash;
+use Symfony\Component\HttpFoundation\Response;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\Admin\StoreUserRequest;
 use App\Http\Requests\Admin\UpdateUserRequest;
 
@@ -15,66 +17,63 @@ use App\Models\Role;
 
 class UsersController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+
     public function index()
     {
-        $users = User::paginate();
+        abort_if(Gate::denies('user_management_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        $users = User::with('roles')->paginate();
 
         return view('admin.users.index',compact('users'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
+
     public function create()
     {
-        $roles = Role::select('id','title')->get()->pluck('title','id');
+        abort_if(Gate::denies('user_management_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        $roles = Role::select('id','title')->get();
 
         return view('admin.users.create',compact('roles'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
+
     public function store(StoreUserRequest $request)
     {
-        $user = User::create($request->validated());
-        $user->roles()->sync($request->input('roles', []));
+        $user = User::create($request->safe()->except(['role_id']));
+        $user->roles()->sync($request->validated('role_id'));
 
         return redirect()->route('admin.users.index');
     }
 
-    /**
-     * Display the specified resource.
-     */
+
     public function show(User $user)
     {
+        abort_if(Gate::denies('user_management_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $user->load('roles');
 
         return view('admin.users.show',compact('user'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
+
     public function edit(User $user)
     {
-        $roles = Role::select('id','title')->get()->pluck('title','id');
+        abort_if(Gate::denies('user_management_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        $roles = Role::select('id','title')->get();
 
         $user->load('roles');
 
         return view('admin.users.edit',compact('roles','user'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
+
     public function update(UpdateUserRequest $request, User $user)
     {
-        $user->update($request->validated());
-        $user->roles()->sync($request->input('roles', []));
+
+        $user->update($request->safe()->except(['role_id']));
+        $user->roles()->sync($request->validated('role_id'));
 
         if(isset($request['password']) && $request['password'] != null){
 
@@ -91,11 +90,11 @@ class UsersController extends Controller
         return redirect()->route('admin.users.show',$user->id)->with('status', 'admin-users-updated');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
+
     public function destroy(User $user)
     {
+        abort_if(Gate::denies('user_management_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $user->delete();
 
         return redirect()->route('admin.users.index');
